covid_api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-cloud-bigquery==3.11.0
pandas==2.0.3
numpy==1.24.3
db-dtypes==1.1.1
functions-framework==3.9.1